pyproject.toml
[project]
name = "task-2"
version = "0.1.0"
description = ""
authors = [
    {name = "sergejlapin",email = "sergey.l37@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
